feat(resolver): per-provider failure summary
